created collections popup component
